move ticks to seperate component
