+++ src/components/RadarChart/RadarChart.tsx
@@ -1,7 +1,8 @@
-import { scaleLinear, select, selectAll } from "d3";
+import { scaleLinear } from "d3";
 import { useEffect } from "react";
 import { useSelector } from "react-redux";
 import { selectTracks } from "../../slices/spotifySlice";
+import Ticks from "./Ticks";
 
 
 const RadarChartComp = () => {
@@ -35,35 +36,16 @@ const RadarChartComp = () => {
         ]
 
         
-    
-        // svg.selectAll("circle")
-        // .data(ticks)
-        // .join(
-        //     enter => enter.append("circle")
-        //         .attr("cx", width / 2)
-        //         .attr("cy", height / 2)
-        //         .attr("fill", "none")
-        //         .attr("stroke", "gray")
-        //         .attr("r", d => radialScale(d))
-        // );
-
-        
         
     },[tracks])
-
-    const svg = select('spider-chart');
 
     const radialScale = scaleLinear()
         .domain([0,100])
         .range([0,2500]);
-    const ticks = [20,40,60,80,100];
+    
     
 
-    useEffect(() => {
-        selectAll(".tick-label")
-        .data(ticks)
-        .text(d => d.toString())
-    },[])
+    
 
 
 
@@ -71,24 +53,11 @@ const RadarChartComp = () => {
         // <svg width={'100%'} viewBox={`0 0 ${width + 200} ${height + 30}`}>
         // somehting like above for responsive ness
         <svg className="spider-chart" width={'100%'} viewBox={`0 0 ${width} ${height}`}>
-            {ticks.map((index,tick) =>
-                <g key={index}>
-                    <text
-                        x={width / 2 + 5}
-                        y={height / 2 - radialScale(tick) - 12}
-                        className='tick-label'  
-                    >
-                    </text> 
-                    <circle
-                        cx={width/2}
-                        cy={height/2}
-                        fill="none"
-                        stroke="gray"
-                        r={radialScale(tick) + 10}
-                    >
-                    </circle>
-                </g>
-            )}
+            <Ticks
+                width={width}
+                height={height}
+                radialScale={radialScale}
+            />
         </svg>
     )
 }
